back_project/requirements/base.txt
amqp==5.0.6
asgiref==3.5.2
billiard==3.6.4.0
boto3==1.18.50
botocore==1.21.50
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.0
channels==4.0.0
charset-normalizer==2.0.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
decorator==4.4.2
defusedxml==0.7.1
dj.choices==0.11.0
Django==4.1.9
django-allauth==0.45.0
django-choices==1.7.2
django-cors-headers==3.8.0
django-filter==2.4.0
django-health-check==3.16.5
django-model-utils==4.2.0
django-rest-auth==0.9.5
django-simple-history==3.0.0
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.2.2
Markdown==3.3.4
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.21.5
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
Pillow==8.3.2
proglog==0.1.9
prompt-toolkit==3.0.23
psutil==5.9.0
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==3.10.0.2
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
sorl-thumbnail==12.7.0
pyfcm
